requirements.txt
asgiref==3.4.1
Babel==2.9.1
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
cryptography==3.4.7
decorator==5.0.9
defusedxml==0.7.1
Django==3.2.5
django-cors-headers==3.7.0
django-crispy-forms==1.12.0
django-extensions==3.1.3
django-phone-field==1.8.1
django-phonenumber-field==5.2.0
django-rest==0.8.7
django-rest-knox==4.1.0
django-waitress==0.1.0
djangoflutterwave==0.3.0
djangorestframework==3.12.4
dynaconf==3.1.4
gunicorn==20.1.0
hvac==0.11.0
idna==2.10
knox==0.1.14
loguru==0.5.3
oauthlib==3.1.1
phonenumbers==8.12.27
pycparser==2.20
PyJWT==2.1.0
pyOpenSSL==20.0.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
shortuuid==1.0.1
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
validators==0.18.2
waitress==1.4.4
Werkzeug==2.0.1
whitenoise==5.2.0
win32-setctime==1.0.3
